separated graph coarsening algorithm into a separate file
